object-array API: use "size_t" for "nr" and "alloc"
Continue the slow march towards guaranteed 64-bit safety by changing
another API's use of "unsigned int" or "unsigned long" to using
"size_t".

In preceding commits we changed most users of the "struct
object_array" to not care about the "nr" at all, and to instead use
the type-agnostic for_each_object_array_entry() helper macro, so
there's very little to change at this point.

Due to 99d60545f87 we can change code in verify_bundle() that kept
track of the "i" and "req_nr" between a "struct string_list"'s "nr"
and our "struct object_array"'s "nr" to use "size_t", as both "nr"
members now share that type.

+++ diff-lib.c
@@ -552,7 +552,7 @@ static int diff_cache(struct rev_info *revs,
 void diff_get_merge_base(const struct rev_info *revs, struct object_id *mb)
 {
 	struct object_array_entry *entry;
-	int i;
+	size_t i;
 	struct commit *mb_child[2] = {0};
 	struct commit_list *merge_bases;
 
@@ -570,7 +570,7 @@ void diff_get_merge_base(const struct rev_info *revs, struct object_id *mb)
 	 * misleading error message.
 	 */
 	if (revs->pending.nr < 1 || revs->pending.nr > 2)
-		BUG("unexpected revs->pending.nr: %d", revs->pending.nr);
+		BUG("unexpected revs->pending.nr: %lu", revs->pending.nr);
 
 	for (i = 0; i < revs->pending.nr; i++)
 		mb_child[i] = lookup_commit_reference(the_repository, &revs->pending.objects[i].item->oid);
